+++ SixTech/API/MapCore/SnapshotManager.swift
