+++ SixTech/View/CharacterCreate/CharacterCreateView.swift
@@ -50,7 +50,7 @@ struct CharacterCreateView: View {
                             
             NavigationLinkView(text: "선택하기"
                                , isdisable: Binding.constant(viewModel.isdisable)
-                               , destination: CameraView(viewModel: CameraViewModel()))
+                               , destination: MainView())
             .simultaneousGesture(TapGesture().onEnded {
                 viewModel.saveUserDefault()
             })
