+++ SixTech/View/ShareResult/ShareResultView.swift
